+++ src/main/java/com/klarna/rest/api/checkout/model/CheckoutAggregatedOrderCartV2.java
@@ -27,7 +27,7 @@ import java.util.List;
 /**
  * CheckoutAggregatedOrderCartV2
  */
-@javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaClientCodegen", date = "2018-11-01T15:17:45.358Z")
+@javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaClientCodegen", date = "2019-01-25T11:03:40.852Z")
 public class CheckoutAggregatedOrderCartV2 {
   @JsonProperty("items")
   private List<CheckoutAggregatedOrderCartItemV2> items = null;