+++ src/main/java/com/klarna/rest/api/checkout/model/CheckoutAggregatedOrderPaymentProviderV2.java
@@ -24,7 +24,7 @@ import io.swagger.annotations.ApiModelProperty;
 /**
  * CheckoutAggregatedOrderPaymentProviderV2
  */
-@javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaClientCodegen", date = "2018-11-01T15:17:45.358Z")
+@javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaClientCodegen", date = "2019-01-25T11:03:40.852Z")
 public class CheckoutAggregatedOrderPaymentProviderV2 {
   @JsonProperty("name")
   private String name = null;