+++ src/main/java/com/klarna/rest/api/checkout/model/CheckoutGui.java
@@ -26,7 +26,7 @@ import java.util.List;
 /**
  * CheckoutGui
  */
-@javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaClientCodegen", date = "2018-11-01T15:17:45.358Z")
+@javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaClientCodegen", date = "2019-01-25T11:03:40.852Z")
 public class CheckoutGui {
   @JsonProperty("options")
   private List<String> options = null;